refactor: change retriable and configurable as mixins
Change retriable and configurable as mixin modules
